New Value mixin
